gdn/requirements.txt
numpy>=1.19.5
pandas>=1.2.1
torch-geometric==1.5.0
torch-scatter==2.1.1
torch-sparse==0.6.17
